requirements.txt
langgraph==0.2.0
google-generativeai==0.7.2
chromadb==0.5.5
pyyaml==6.0.1
sentence-transformers==2.7.0
langchain-community>=0.2.0
langchain-core==0.2.43
langchain
transformers
torch
datasets